+++ server/config/config.go
@@ -35,16 +35,19 @@ const (
 	DefaultGRPCAddress = "0.0.0.0:9900"
 
 	// DefaultJSONRPCAddress is the default address the JSON-RPC server binds to.
-	DefaultJSONRPCAddress = "0.0.0.0:8545"
+	DefaultJSONRPCAddress = "127.0.0.1:8545"
 
 	// DefaultJSONRPCWsAddress is the default address the JSON-RPC WebSocket server binds to.
-	DefaultJSONRPCWsAddress = "0.0.0.0:8546"
+	DefaultJSONRPCWsAddress = "127.0.0.1:8546"
 
 	// DefaultJsonRPCMetricsAddress is the default address the JSON-RPC Metrics server binds to.
-	DefaultJSONRPCMetricsAddress = "0.0.0.0:6065"
+	DefaultJSONRPCMetricsAddress = "127.0.0.1:6065"
 
 	// DefaultEVMTracer is the default vm.Tracer type
 	DefaultEVMTracer = ""
+
+	// DefaultFixRevertGasRefundHeight is the default height at which to overwrite gas refund
+	DefaultFixRevertGasRefundHeight = 0
 
 	DefaultMaxTxGasWanted = 0
 
@@ -59,14 +62,17 @@ const (
 	DefaultBlockRangeCap int32 = 10000
 
 	DefaultEVMTimeout = 5 * time.Second
+
 	// default 1.0 eth
 	DefaultTxFeeCap float64 = 1.0
 
 	DefaultHTTPTimeout = 30 * time.Second
 
 	DefaultHTTPIdleTimeout = 120 * time.Second
+
 	// DefaultAllowUnprotectedTxs value is false
 	DefaultAllowUnprotectedTxs = false
+
 	// DefaultMaxOpenConnections represents the amount of open connections (unlimited = 0)
 	DefaultMaxOpenConnections = 0
 )
@@ -130,6 +136,8 @@ type JSONRPCConfig struct {
 	EnableIndexer bool `mapstructure:"enable-indexer"`
 	// MetricsAddress defines the metrics server to listen on
 	MetricsAddress string `mapstructure:"metrics-address"`
+	// FixRevertGasRefundHeight defines the upgrade height for fix of revert gas refund logic when transaction reverted
+	FixRevertGasRefundHeight int64 `mapstructure:"fix-revert-gas-refund-height"`
 }
 
 // TLSConfig defines the certificate and matching private key for the server.
@@ -215,23 +223,24 @@ func GetAPINamespaces() []string {
 // DefaultJSONRPCConfig returns an EVM config with the JSON-RPC API enabled by default
 func DefaultJSONRPCConfig() *JSONRPCConfig {
 	return &JSONRPCConfig{
-		Enable:              true,
-		API:                 GetDefaultAPINamespaces(),
-		Address:             DefaultJSONRPCAddress,
-		WsAddress:           DefaultJSONRPCWsAddress,
-		GasCap:              DefaultGasCap,
-		EVMTimeout:          DefaultEVMTimeout,
-		TxFeeCap:            DefaultTxFeeCap,
-		FilterCap:           DefaultFilterCap,
-		FeeHistoryCap:       DefaultFeeHistoryCap,
-		BlockRangeCap:       DefaultBlockRangeCap,
-		LogsCap:             DefaultLogsCap,
-		HTTPTimeout:         DefaultHTTPTimeout,
-		HTTPIdleTimeout:     DefaultHTTPIdleTimeout,
-		AllowUnprotectedTxs: DefaultAllowUnprotectedTxs,
-		MaxOpenConnections:  DefaultMaxOpenConnections,
-		EnableIndexer:       false,
-		MetricsAddress:      DefaultJSONRPCMetricsAddress,
+		Enable:                   true,
+		API:                      GetDefaultAPINamespaces(),
+		Address:                  DefaultJSONRPCAddress,
+		WsAddress:                DefaultJSONRPCWsAddress,
+		GasCap:                   DefaultGasCap,
+		EVMTimeout:               DefaultEVMTimeout,
+		TxFeeCap:                 DefaultTxFeeCap,
+		FilterCap:                DefaultFilterCap,
+		FeeHistoryCap:            DefaultFeeHistoryCap,
+		BlockRangeCap:            DefaultBlockRangeCap,
+		LogsCap:                  DefaultLogsCap,
+		HTTPTimeout:              DefaultHTTPTimeout,
+		HTTPIdleTimeout:          DefaultHTTPIdleTimeout,
+		AllowUnprotectedTxs:      DefaultAllowUnprotectedTxs,
+		MaxOpenConnections:       DefaultMaxOpenConnections,
+		EnableIndexer:            false,
+		MetricsAddress:           DefaultJSONRPCMetricsAddress,
+		FixRevertGasRefundHeight: DefaultFixRevertGasRefundHeight,
 	}
 }
 
@@ -325,22 +334,23 @@ func GetConfig(v *viper.Viper) (Config, error) {
 			MaxTxGasWanted: v.GetUint64("evm.max-tx-gas-wanted"),
 		},
 		JSONRPC: JSONRPCConfig{
-			Enable:             v.GetBool("json-rpc.enable"),
-			API:                v.GetStringSlice("json-rpc.api"),
-			Address:            v.GetString("json-rpc.address"),
-			WsAddress:          v.GetString("json-rpc.ws-address"),
-			GasCap:             v.GetUint64("json-rpc.gas-cap"),
-			FilterCap:          v.GetInt32("json-rpc.filter-cap"),
-			FeeHistoryCap:      v.GetInt32("json-rpc.feehistory-cap"),
-			TxFeeCap:           v.GetFloat64("json-rpc.txfee-cap"),
-			EVMTimeout:         v.GetDuration("json-rpc.evm-timeout"),
-			LogsCap:            v.GetInt32("json-rpc.logs-cap"),
-			BlockRangeCap:      v.GetInt32("json-rpc.block-range-cap"),
-			HTTPTimeout:        v.GetDuration("json-rpc.http-timeout"),
-			HTTPIdleTimeout:    v.GetDuration("json-rpc.http-idle-timeout"),
-			MaxOpenConnections: v.GetInt("json-rpc.max-open-connections"),
-			EnableIndexer:      v.GetBool("json-rpc.enable-indexer"),
-			MetricsAddress:     v.GetString("json-rpc.metrics-address"),
+			Enable:                   v.GetBool("json-rpc.enable"),
+			API:                      v.GetStringSlice("json-rpc.api"),
+			Address:                  v.GetString("json-rpc.address"),
+			WsAddress:                v.GetString("json-rpc.ws-address"),
+			GasCap:                   v.GetUint64("json-rpc.gas-cap"),
+			FilterCap:                v.GetInt32("json-rpc.filter-cap"),
+			FeeHistoryCap:            v.GetInt32("json-rpc.feehistory-cap"),
+			TxFeeCap:                 v.GetFloat64("json-rpc.txfee-cap"),
+			EVMTimeout:               v.GetDuration("json-rpc.evm-timeout"),
+			LogsCap:                  v.GetInt32("json-rpc.logs-cap"),
+			BlockRangeCap:            v.GetInt32("json-rpc.block-range-cap"),
+			HTTPTimeout:              v.GetDuration("json-rpc.http-timeout"),
+			HTTPIdleTimeout:          v.GetDuration("json-rpc.http-idle-timeout"),
+			MaxOpenConnections:       v.GetInt("json-rpc.max-open-connections"),
+			EnableIndexer:            v.GetBool("json-rpc.enable-indexer"),
+			MetricsAddress:           v.GetString("json-rpc.metrics-address"),
+			FixRevertGasRefundHeight: v.GetInt64("json-rpc.fix-revert-gas-refund-height"),
 		},
 		TLS: TLSConfig{
 			CertificatePath: v.GetString("tls.certificate-path"),