+++ x/evm/keeper/params.go
@@ -23,10 +23,9 @@ import (
 // GetParams returns the total set of evm parameters.
 func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
 	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.ParamsKey)
+	bz := store.Get(types.KeyPrefixParams)
 	if len(bz) == 0 {
-		params = k.GetLegacyParams(ctx)
-		return
+		return k.GetLegacyParams(ctx)
 	}
 	k.cdc.MustUnmarshal(bz, &params)
 	return
@@ -44,7 +43,7 @@ func (k Keeper) SetParams(ctx sdk.Context, params types.Params) error {
 		return err
 	}
 
-	store.Set(types.ParamsKey, bz)
+	store.Set(types.KeyPrefixParams, bz)
 	return nil
 }