+++ x/evm/migrations/v4/migrate_test.go
@@ -3,14 +3,15 @@ package v4_test
 import (
 	"testing"
 
-	"github.com/evmos/ethermint/x/evm/types"
+	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/evmos/ethermint/app"
 	"github.com/evmos/ethermint/encoding"
 	v4 "github.com/evmos/ethermint/x/evm/migrations/v4"
-	"github.com/stretchr/testify/require"
+	"github.com/evmos/ethermint/x/evm/types"
 )
 
 type mockSubspace struct {
@@ -30,14 +31,14 @@ func TestMigrate(t *testing.T) {
 	cdc := encCfg.Codec
 
 	storeKey := sdk.NewKVStoreKey(types.ModuleName)
-	tKey := sdk.NewTransientStoreKey("transient_test")
+	tKey := sdk.NewTransientStoreKey(types.TransientKey)
 	ctx := testutil.DefaultContext(storeKey, tKey)
 	kvStore := ctx.KVStore(storeKey)
 
 	legacySubspace := newMockSubspace(types.DefaultParams())
 	require.NoError(t, v4.MigrateStore(ctx, storeKey, legacySubspace, cdc))
 
-	paramsBz := kvStore.Get(types.ParamsKey)
+	paramsBz := kvStore.Get(types.KeyPrefixParams)
 	var params types.Params
 	cdc.MustUnmarshal(paramsBz, &params)